Filtro para busqueda de alumnos
***Agregar CSS ***

+++ client/src/pages/ShowAlu.jsx
@@ -16,38 +16,69 @@ export const ShowEstudiante = () => {
     
         //Mostrar Todos los estudiantes de la Base de Datos
         const [estudiantes, setEstudiantes] = React.useState([]);
+        const [busqueda, setBusqueda] = React.useState("");
+        const [tablaUsuarios, setTablaUsuarios] = React.useState([]);
     
         React.useEffect(() => {
             fetch("http://localhost:5000/estudiante")
                 .then((response) => response.json())
-                .then((data) => setEstudiantes(data));
+                .then((data) => {
+                    setEstudiantes(data)
+                    setTablaUsuarios(data)
+                });
+                
         }, []);
 
         function handleDelete(id) {
-            fetch(`http://localhost:5000/estudiante/${id}`, {
+            if(window.confirm("¿Está seguro de eliminar este estudiante?")) {
+                fetch(`http://localhost:5000/estudiante/${id}`, {
                 method: "DELETE",
             }).then((response) => {
                 if (response.ok) {
                     alert("Estudiante Eliminado");
-                    window.location.reload();
+                    // Actualizar la lista de estudiantes
+                    fetch("http://localhost:5000/estudiante")
+                        .then((response) => response.json())
+                        .then((data) => setEstudiantes(data));
                 }
             });
+            }
         }
 
         function handleEdit(id) {
-            fetch(`http://localhost:5000/estudiante/${id}`, {
-                method: "PUT",
-            }).then((response) => {
-                if (response.ok) {
-                    alert("Estudiante Editado");
-                    window.location.reload();
+            
+        }
+        //handlechange para buscar por nombre o apellido
+        function handleChange(e) {
+            setBusqueda(e.target.value);
+            filtrar(e.target.value);
+        }
+        const filtrar = (terminoBusqueda) => {
+            var resultadosBusqueda = tablaUsuarios.filter((elemento) => {
+                if (
+                    elemento.nombre.toLowerCase().includes(terminoBusqueda.toLowerCase()) ||
+                    elemento.apellido.toLowerCase().includes(terminoBusqueda.toLowerCase())
+                ) {
+                    return elemento;
                 }
             });
-        }
+            setEstudiantes(resultadosBusqueda);
+        };
+
 
     return (
+        <>
+        {/* Tabla de Estudiantes */}
         <div class="overflow-x-auto">
             <h1>Estudiantes</h1>
+            {/* Buscador de Estudiante */}
+            <div className="container">
+                <input type="text" 
+                placeholder="Buscar Estudiante por Nombre o Apellido" 
+                onChange={handleChange} 
+                value={busqueda} 
+                style={{color : "black"}}/>
+            </div>
             <table class="table w-full">
                 <thead>
                     <tr>
@@ -55,6 +86,7 @@ export const ShowEstudiante = () => {
                         <th>Apellido</th>
                         <th>Rut</th>
                         <th>Email</th>
+                        <th>Contraseña</th>
                         <th>Año de Ingreso</th>
                     </tr>
                 </thead>
@@ -65,15 +97,20 @@ export const ShowEstudiante = () => {
                             <td>{estudiante.apellido}</td>
                             <td>{estudiante.rut}</td>
                             <td>{estudiante.email}</td>
+                            <td>{estudiante.password}</td>
                             <td>{estudiante.ingreso}</td>
                             <td>
                                 <button onClick={() => handleDelete(estudiante._id)}>Eliminar</button>
+                            </td>
+                            <td>
+                                <button onClick={() => handleEdit(estudiante._id)}>Editar</button>
                             </td>
                         </tr>
                     ))}
                 </tbody>
             </table>
         </div>
+        </>
     );
 };
